Section 9 Completed

+++ backend/routes/posts.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const multer = require('multer');
 
-const router = express.Router();
-
 const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
+
+const router = express.Router();
 
 const MIME_TYPE_MAP = {
   'image/png': 'png',
@@ -27,42 +28,54 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post('', multer({storage}).single('image'), (req, res, next) => {
-  const url = req.protocol + '://' + req.get('host');
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
-  });
-  post.save()
-    .then(createdPost => {
-      res.status(201).json({
-        message: 'Post added successfully!',
-        post: {
-          ...createdPost,
-          id: createdPost._id
-        }
-      });
-    });
-});
-
-router.put('/:id', multer({storage}).single('image'), (req, res, next) => {
-  let imagePath = req.body.imagePath;
-  if (req.file) {
+router.post(
+  '',
+  checkAuth,
+  multer({storage}).single('image'),
+  (req, res, next) => {
     const url = req.protocol + '://' + req.get('host');
-    imagePath = url + '/images/' + req.file.filename;
-  }
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath
+    const post = new Post({
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
+    });
+    post.save()
+      .then(createdPost => {
+        res.status(201).json({
+          message: 'Post added successfully!',
+          post: {
+            ...createdPost,
+            id: createdPost._id
+          }
+        });
+      });
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Update successful!"});
-  })
-});
+
+router.put('/:id',
+  checkAuth,
+  multer({storage}).single('image'),
+  (req, res, next) => {
+    let imagePath = req.body.imagePath;
+    if (req.file) {
+      const url = req.protocol + '://' + req.get('host');
+      imagePath = url + '/images/' + req.file.filename;
+    }
+    const post = new Post({
+      _id: req.body.id,
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
+    });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if (result.modifiedCount > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({message: "Not Authorized!"});
+      }
+    })
+  });
 
 router.get('/:id', (req, res, next) => {
   Post.findById(req.params.id).then(post => {
@@ -98,12 +111,17 @@ router.get('', (req, res, next) => {
     });
 });
 
-router.delete('/:id', (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
-    .then(result => {
-      console.log(result);
-      res.status(200).json({message: "Post deleted!"});
-    });
-});
+router.delete('/:id',
+  checkAuth,
+  (req, res, next) => {
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
+      .then(result => {
+        if (result.deletedCount > 0) {
+          res.status(200).json({message: "Post deleted!"});
+        } else {
+          res.status(401).json({message: "Not Authorized!"});
+        }
+      });
+  });
 
 module.exports = router;
